backend/TravelAgencySite/requirements.txt
absl-py==1.3.0
anyio==3.6.2
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asgiref==3.7.2
asttokens==2.0.8
astunparse==1.6.3
async-timeout==4.0.2
attrs==23.1.0
autobahn==23.1.2
Automat==22.10.0
backcall==0.2.0
backports.zoneinfo==0.2.1
beautifulsoup4==4.9.1
bleach==5.0.1
cachetools==5.2.0
certifi==2023.5.7
cffi==1.15.1
channels==4.0.0
channels-redis==4.1.0
chardet==4.0.0
charset-normalizer==3.1.0
click==8.1.3
cmdstanpy==0.9.5
colorama==0.4.6
constantly==15.1.0
contourpy==1.0.5
convertdate==2.4.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==40.0.2
cssselect==1.2.0
cycler==0.11.0
Cython==0.29.32
daphne==4.0.0
debugpy==1.6.3
decorator==5.1.1
defusedxml==0.7.1
Deprecated==1.2.13
dj-database-url==2.0.0
Django==4.2.1
django-admin-autocomplete-filter==0.7.1
django-allauth==0.54.0
django-cors-headers==4.0.0
django-environ==0.10.0
django-extensions==3.2.3
django-filter==23.2
django-rest-auth==0.9.5
django-rest-swagger==2.2.0
django-simple-history==3.3.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
drf-yasg==1.21.5
entrypoints==0.4
ephem==4.1.3
executing==1.1.1
fastjsonschema==2.16.2
filelock==3.8.0
flatbuffers==22.10.26
fonttools==4.38.0
gast==0.4.0
google-auth==2.14.1
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
graphviz==0.20.1
grpcio==1.50.0
gunicorn==20.1.0
h5py==3.7.0
hijri-converter==2.2.4
holidays==0.16
hyperlink==21.0.0
idna==3.4
importlib-metadata==5.0.0
importlib-resources==5.10.0
incremental==22.10.0
inflection==0.5.1
ipykernel==6.16.2
ipython==8.5.0
ipython-genutils==0.2.0
itemadapter==0.7.0
itemloaders==1.0.6
itsdangerous==2.1.2
itypes==1.2.0
jdatetime==4.1.0
jedi==0.18.1
Jinja2==3.1.2
jmespath==1.0.1
joblib==1.2.0
jsonschema==4.17.3
Keras-Preprocessing==1.1.2
kiwisolver==1.4.4
korean-lunar-calendar==0.3.1
libclang==14.0.6
lightgbm==3.3.3
limits==3.3.1
LunarCalendar==0.0.9
lxml==4.9.1
Markdown==3.4.1
markdown-it-py==2.2.0
MarkupSafe==2.1.3
matplotlib-inline==0.1.6
mdurl==0.1.2
mistune==2.0.4
mlxtend==0.21.0
msgpack==1.0.5
nbclassic==0.4.5
nbclient==0.7.0
nbconvert==7.2.2
nbformat==5.7.0
nest-asyncio==1.5.6
oauthlib==3.2.2
openapi-codec==1.3.2
opt-einsum==3.3.0
ordered-set==4.1.0
packaging==23.1
pandocfilters==1.5.0
parsel==1.7.0
parso==0.8.3
patsy==0.5.3
pickleshare==0.7.5
Pillow==9.5.0
pkgutil_resolve_name==1.3.10
prometheus-client==0.15.0
prompt-toolkit==3.0.31
Protego==0.2.1
protobuf==3.19.6
psutil==5.9.3
psycopg2==2.9.6
psycopg2-binary==2.9.6
pure-eval==0.2.2
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
PyDispatcher==2.0.6
pydot==1.4.2
Pygments==2.13.0
PyJWT==2.7.0
PyMeeus==0.5.11
pyOpenSSL==23.1.1
pyparsing==3.0.9
pyrsistent==0.19.3
pystan==2.19.1.1
python-dateutil==2.8.2
python-decouple==3.8
python-sec==0.1.6
python3-openid==3.2.0
pytz==2023.3
pywin32==304
pywinpty==2.0.9
PyYAML==6.0
pyzmq==24.0.1
queuelib==1.6.2
redis==4.5.5
requests==2.31.0
requests-file==1.5.1
requests-oauthlib==1.3.1
rich==13.3.3
rsa==4.9
ruamel.yaml==0.17.31
ruamel.yaml.clib==0.2.7
scikit-learn==1.1.3
Scrapy==2.7.1
sec==0.3.0
Send2Trash==1.8.0
service-identity==21.1.0
setuptools-git==1.2
simplejson==3.19.1
six==1.16.0
sniffio==1.3.0
soupsieve==2.3.2.post1
sqlparse==0.4.4
stack-data==0.5.1
statsmodels==0.13.2
swagger-spec-validator==3.0.3
tensorboard==2.10.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
termcolor==2.1.0
terminado==0.17.0
threadpoolctl==3.1.0
tinycss2==1.2.1
tldextract==3.4.0
tornado==6.2
tqdm==4.64.1
traitlets==5.5.0
Twisted==22.10.0
twisted-iocpsupport==1.0.2
txaio==23.1.1
typing_extensions==4.6.2
tzdata==2023.3
uritemplate==4.1.1
urllib3==2.0.2
w3lib==2.0.1
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==1.4.1
Werkzeug==2.2.2
whitenoise==6.5.0
wrapt==1.14.1
xgboost==1.6.2
yellowbrick==1.5
zipp==3.10.0
zope.interface==6.0
